+++ xcodeproj/internal/bazel_integration_files/generate_index_build_bazel_dependencies.sh
@@ -63,7 +63,13 @@ source "$BAZEL_INTEGRATION_DIR/bazel_build.sh"
 
 # Import indexes
 if [ -n "${indexstores_filelists:-}" ]; then
-  "$BAZEL_INTEGRATION_DIR/import_indexstores" \
-    "$PROJECT_DIR" \
-    "${indexstores_filelists[@]/#/$BAZEL_OUT/}"
+  if [[ "${BAZEL_SEPARATE_INDEXBUILD_OUTPUT_BASE:-}" == "YES" ]]; then
+    "$BAZEL_INTEGRATION_DIR/import_indexstores" \
+      "$INDEXING_PROJECT_DIR__NO" \
+      "${indexstores_filelists[@]/#/$output_path/}"
+  else
+    "$BAZEL_INTEGRATION_DIR/import_indexstores" \
+      "$PROJECT_DIR" \
+      "${indexstores_filelists[@]/#/$BAZEL_OUT/}"
+  fi
 fi