+++ xcodeproj/internal/templates/generate_bazel_dependencies.sh
@@ -192,8 +192,15 @@ done
 
 # Import indexes
 if [ -n "${indexstores_filelists:-}" ]; then
-  "$BAZEL_INTEGRATION_DIR/import_indexstores" \
-    "$PROJECT_DIR" \
-    "${indexstores_filelists[@]/#/$BAZEL_OUT/}" \
-    >"$log_dir/import_indexstores.async.log" 2>&1 &
+  if [[ "${BAZEL_SEPARATE_INDEXBUILD_OUTPUT_BASE:-}" == "YES" ]]; then
+    "$BAZEL_INTEGRATION_DIR/import_indexstores" \
+      "$INDEXING_PROJECT_DIR__NO" \
+      "${indexstores_filelists[@]/#/$output_path/}" \
+      >"$log_dir/import_indexstores.async.log" 2>&1 &
+  else
+    "$BAZEL_INTEGRATION_DIR/import_indexstores" \
+      "$PROJECT_DIR" \
+      "${indexstores_filelists[@]/#/$BAZEL_OUT/}" \
+      >"$log_dir/import_indexstores.async.log" 2>&1 &
+  fi
 fi